Architecture refactoring.
