+++ src/app/characters/characters.module.ts
@@ -2,11 +2,10 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CharactersComponent } from './pages/characters/characters.component';
 import { SingleCharacterComponent } from './pages/single-character/single-character.component';
-import { CharacterListComponent } from './pages/characters/character-list/character-list.component';
+import { CharacterListComponent } from './components/character-list/character-list.component';
 import {RouterLink} from "@angular/router";
 import {SharedModule} from "../shared/shared.module";
-import { CharacterCardComponent } from './pages/characters/character-list/character-card/character-card.component';
-import { CharactersContainerComponent } from './pages/characters/characters-container/characters-container.component';
+import { CharacterCardComponent } from './components/character-card/character-card.component';
 
 
 
@@ -16,7 +15,6 @@ import { CharactersContainerComponent } from './pages/characters/characters-cont
     SingleCharacterComponent,
     CharacterListComponent,
     CharacterCardComponent,
-    CharactersContainerComponent,
   ],
   imports: [
     CommonModule,
