+++ src/app/characters/components/character-card/character-card.component.css
