+++ src/app/characters/components/character-list/character-list.component.html
