+++ src/app/characters/models/charactersPageModel.ts
@@ -1,0 +1,15 @@
+import {PaginatedModelState} from "../../shared/models/paginatedModelState";
+import {Character} from "../../shared/models/character";
+
+export type CharactersState = PaginatedModelState<Character[]>;
+
+export const charactersPageInitialState: CharactersState = {
+  isLoading: false,
+  isLoaded: false,
+  paginationInfo: {
+    hasPrev: false,
+    hasNext: false,
+  },
+  data: [],
+  error: null,
+};
