--- src/app/characters/pages/characters/characters-container/characters-container.component.html
@@ -1,1 +1,0 @@
-<p>characters-container works!</p>
