--- src/app/characters/pages/characters/characters-container/characters-container.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-characters-container',
-  templateUrl: './characters-container.component.html',
-  styleUrls: ['./characters-container.component.css']
-})
-export class CharactersContainerComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
