+++ src/app/shared/components/input-filter/input-filter.component.css
@@ -1,5 +1,6 @@
 .filter-form {
   margin: 3rem 0 0 0;
+  display: inline-block;
   width: 100%;
   outline: 0;
   border: #a6a3a3 solid 1px;
@@ -18,3 +19,8 @@
   border-color: black  ;
 }
 
+.form-button {
+  display: inline-block;
+  width: 100%;
+  border-radius: 1rem;
+}
