+++ src/app/shared/components/input-filter/input-filter.component.html
@@ -1,1 +1,2 @@
-<input #input class="center filter-form" placeholder="Filter by name..." type="text">
+<input #input  class="filter-form" placeholder="Filter by name..." type="text">
+
